Move http object to a separate file
