+++ src/components/App/Layout/Layout.jsx
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import { translate } from 'react-i18next';
 import Header from 'components/App/Header/Header';
 import Footer from 'components/App/Footer/Footer';
+import http from './http.helper';
 
 class Layout extends Component {
   constructor(props) {
@@ -9,39 +10,6 @@ class Layout extends Component {
     this.state = {
       user: {
         token: localStorage.getItem('userToken'),
-      },
-    };
-    this.http = {
-      get(url, params, options) {
-        const parseMethod = options ? options.parse : 'json';
-        let getUrl = url.replace('/api/', `${process.env.REACT_APP_API}/`);
-
-        if (params) {
-          const getParams = new URLSearchParams(Object.entries(params));
-          getUrl = `${getUrl}?${getParams}`;
-        }
-
-        if (parseMethod === 'none') return fetch(getUrl);
-
-        return fetch(getUrl)
-          .then(response => response[parseMethod]());
-      },
-
-      post(url, body, options) {
-        const parseMethod = options ? options.parse : 'json';
-        const postUrl = url.replace('/api/', `${process.env.REACT_APP_API}/`);
-        const myInit = {
-          body: JSON.stringify(body),
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          method: 'POST',
-        };
-
-        if (parseMethod === 'none') return fetch(postUrl, myInit);
-
-        return fetch(postUrl, myInit)
-          .then(response => response[parseMethod]());
       },
     };
     this.setUserToken = this.setUserToken.bind(this);
@@ -55,7 +23,7 @@ class Layout extends Component {
     const { children } = this.props;
     const childProps = {
       setUserToken: this.setUserToken,
-      http: this.http,
+      http,
     };
     const childrenWithProps = React.Children.map(children, (child) => {
       if (React.isValidElement(child)) return React.cloneElement(child, childProps);
